+++ bin/blocks/accordion.sh
@@ -2,7 +2,7 @@
 # @Author: Roni Laukkarinen
 # @Date:   2022-02-10 10:44:02
 # @Last Modified by:   Roni Laukkarinen
-# @Last Modified time: 2022-02-10 16:10:01
+# @Last Modified time: 2022-08-18 15:07:05
 
 # // New files/Dependencies (this file will install them):
 # // ├── sass/gutenberg/blocks/_accordion.scss (automatic from get-block.sh)
@@ -21,6 +21,7 @@
 # Block specific variables
 export BLOCK_ACF_JSON_FILE="group_61d8052f4ea71.json"
 export BLOCK_ACF_JSON_PATH="${AIRBLOCKS_THEME_PATH}/acf-json/${BLOCK_ACF_JSON_FILE}"
+export BLOCK_UI_TITLE="Haitari"
 
 # JavaScript dependencies
 cp -nv ${AIRBLOCKS_THEME_PATH}/js/src/modules/accordion.js ${PROJECT_THEME_PATH}/js/src/modules/accordion.js
@@ -42,11 +43,3 @@ mv ${PROJECT_THEME_PATH}/sass/gutenberg-editor-styles-with-changes.scss ${PROJEC
 # Other SVG icons needed by this block
 cp -nv ${AIRBLOCKS_THEME_PATH}/svg/minus.svg ${PROJECT_THEME_PATH}/svg/
 cp -nv ${AIRBLOCKS_THEME_PATH}/svg/plus.svg ${PROJECT_THEME_PATH}/svg/
-
-# Register ACF block in functions.php
-# Please note: The title of the block will be translated in localization.sh if en is selected
-sed -e "/\'acf_blocks\' \=\> \[/a\\
-      [|\
-       'name' => 'accordion',|\
-       'title' => 'Haitari',|\
-      ],\\" < ${PROJECT_THEME_PATH}/functions.php | tr '|' '\n' > ${PROJECT_THEME_PATH}/tmpfile