+++ bin/blocks/hero-full-height.sh
@@ -2,7 +2,7 @@
 # @Author: Roni Laukkarinen
 # @Date:   2021-11-23 18:23:24
 # @Last Modified by:   Roni Laukkarinen
-# @Last Modified time: 2022-02-10 12:18:38
+# @Last Modified time: 2022-08-18 15:16:53
 
 # // New files/Dependencies (this file will install them):
 # // ├── sass/gutenberg/blocks/_hero-full-height.scss (automatic from get-block.sh)
@@ -21,6 +21,7 @@
 # Block specific variables
 export BLOCK_ACF_JSON_FILE="group_618bb10852874.json"
 export BLOCK_ACF_JSON_PATH="${AIRBLOCKS_THEME_PATH}/acf-json/${BLOCK_ACF_JSON_FILE}"
+export BLOCK_UI_TITLE="Sivun korkea yläosa"
 
 # JavaScript dependencies
 cp -nv ${AIRBLOCKS_THEME_PATH}/js/src/modules/100vh.js ${PROJECT_THEME_PATH}/js/src/modules/100vh.js
@@ -44,11 +45,3 @@ mv ${PROJECT_THEME_PATH}/sass/components/_components_with_changes.scss ${PROJECT
 
 # Other SVG icons needed by this block
 # cp -nv ${AIRBLOCKS_THEME_PATH}/svg/play.svg ${PROJECT_THEME_PATH}/svg/
-
-# Register ACF block in functions.php
-# Please note: The title of the block will be translated in localization.sh if en is selected
-sed -e "/\'acf_blocks\' \=\> \[/a\\
-      [|\
-       'name' => 'hero-full-height',|\
-       'title' => 'Sivun korkea yläosa',|\
-      ],\\" < ${PROJECT_THEME_PATH}/functions.php | tr '|' '\n' > ${PROJECT_THEME_PATH}/tmpfile