+++ bin/tasks/post-block.sh
@@ -5,6 +5,23 @@
 # @Last Modified time: 2022-03-09 12:06:34
 # Tasks after functions.php has been updated
 
+# Register ACF block in functions.php
+# Please note: The title of the block will be translated in localization.sh if en is selected
+if [[ ${BLOCK_NAME} = "form" ]]; then
+  sed -e "/\'acf_blocks\' \=\> \[/a\\
+        [|\
+         'name' => '${BLOCK_NAME}',|\
+         'title' => '${BLOCK_UI_TITLE}',|\
+         'prevent_cache' => true,|\
+        ],\\" < ${PROJECT_THEME_PATH}/functions.php | tr '|' '\n' > ${PROJECT_THEME_PATH}/tmpfile
+else
+  sed -e "/\'acf_blocks\' \=\> \[/a\\
+        [§\
+         'name' => '${BLOCK_NAME}',§\
+         'title' => '${BLOCK_UI_TITLE}',§\
+        ],\\" < ${PROJECT_THEME_PATH}/functions.php | tr '§' '\n' > ${PROJECT_THEME_PATH}/tmpfile
+fi
+
 # Let's go to the project directory
 cd ${PROJECTS_HOME}/${PROJECT_NAME}